RM 2523: Various updates
- If model resolution is changed, body will get positioned correctly
- Kernel ingestor import mechanism now smoother
- Available kernels gets refreshed when ingestor frame loads
- Update to SpiceBodyOperator to handle system configs better
- Temporary workaround for DART frame updates and our current lookup
system

+++ src/edu/jhuapl/sbmt/pointing/spice/ingestion/model/SpiceKernelIngestor.java
@@ -49,7 +49,10 @@ public class SpiceKernelIngestor
         {
         	currentFile++;
         	if (listener != null) listener.percentageLoaded((int)(100*(currentFile/numberOfFiles)));
-        	FileUtils.copyDirectoryToDirectory(file.getParentFile(), newDirectory);
+        	if (!(new File(newDirectory, file.getParentFile().getName()).exists()))
+        	{
+        		FileUtils.copyDirectoryToDirectory(file.getParentFile(), newDirectory);
+        	}
         }
         if (listener != null) listener.percentageLoaded(0);
         //update the PATH_VALUES line in the metakernel to match the new directory
